Module Octez_smart_rollup_node_alpha

module Arith_pvm : sig ... end

module Batcher_constants : sig ... end

module Context_wrapper : sig ... end

module Daemon_helpers : sig ... end

module Dal_pages_request : sig ... end

module Dal_slots_tracker : sig ... end

The rollup node keeps the list of dal slots for each block it needs to process. This is to determine whether the inbox for a given block will need to be retrieved from the block operations, or from the data availability layer after lag levels have passed and the slot for the block has been declared available.

module Dal_slots_tracker_event : sig ... end

module Fueled_pvm : sig ... end

module Inbox : sig ... end

module Inbox_event : sig ... end

module Layer1_helpers : sig ... end

module Outbox : sig ... end

module Pvm : sig ... end

module Pvm_plugin : sig ... end

module Pvm_rpc : sig ... end

module Pvm_sig : sig ... end

module RPC_directory : sig ... end

module Refutation_game_helpers : sig ... end

module Reveals : sig ... end

module Riscv_pvm : sig ... end

module Rollup_node_plugin : sig ... end

module Sc_rollup_injector : sig ... end

module Sc_rollup_node_errors : sig ... end

module Wasm_2_0_0_pvm : sig ... end

module Wasm_2_0_0_rpc : sig ... end