Module Impl.State

State storage for this PVM.

type value = state

val empty : unit -> state

empty () is the empty state.

val find : ('a, repo, tree) Tezos_layer2_store.Context_sigs.t -> state option Lwt.t

find context returns the PVM state stored in the context, if any.

val lookup : state -> string list -> bytes option Lwt.t

lookup state path returns the data stored for the path path in the PVM state state.

val set : ('a, repo, tree) Tezos_layer2_store.Context_sigs.t -> state -> ('a, repo, tree) Tezos_layer2_store.Context_sigs.t Lwt.t

set context state saves the PVM state state in the context and returns the updated context. Note: set does not perform any write on disk, this information must be committed using Context.commit.