Module Octez_smart_rollup_node_alpha.Batcher_constants

val message_size_limit : int

Maximum size of an L2 message allowed by the prototcol. Is Protocol.Constants_repr.sc_rollup_message_size_limit.

val protocol_max_batch_size : int

Maximum size in bytes of an batch of L2 messages that can fit in an operation on L1. It is protocol dependent.