Module Octez_smart_rollup_node_alpha.Dal_pages_request

Access DAL slots and pages content.

This module is a wrapper on top of Store.Dal_slot_pages module to access DAL slots and pages' data that have been previously fetched by the rollup node.

val slot_pages : Octez_smart_rollup.Rollup_constants.dal_constants -> dal_activation_level: Tezos_protocol_alpha.Protocol.Alpha_context.Raw_level.t option -> inbox_level:int32 -> _ Octez_smart_rollup_node.Node_context.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Dal.slot_id -> dal_attested_slots_validity_lag:int -> Tezos_protocol_alpha.Protocol.Alpha_context.Dal.Page.content list option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the pages' content of the given slot ID's from the store.

The function returns Dal_slot_not_found_in_store if no entry is found in the store for the given ID (i.e. no page is registered with or without content).

If the returned value is Some pages, the slot whose ID is given is supposed to be confirmed and pages correspond to the pages of the slot. Otherwise None is returned.

The function relies on Store.Dal_slot_pages's invariants to guarantee that:

the pages are returned in increasing order w.r.t. their indexes in the slot;

the size of the list, in case it is not empty, is equal to the expected number of pages in a slot.

dal_attestation_lag is used to retrieve the correct entry in store.

val page_content : Octez_smart_rollup.Rollup_constants.dal_constants -> dal_activation_level: Tezos_protocol_alpha.Protocol.Alpha_context.Raw_level.t option -> inbox_level:int32 -> _ Octez_smart_rollup_node.Node_context.t -> attestation_threshold_percent:int option -> restricted_commitments_publishers: Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t list option -> Tezos_protocol_alpha.Protocol.Alpha_context.Dal.Page.t -> dal_attested_slots_validity_lag:int -> Tezos_protocol_alpha.Protocol.Alpha_context.Dal.Page.content option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the content of the page identified by the given ID from the store.

This function checks the attestation status of the page's slot by inspecting the skip list cells stored on L1 at the slot's attested level. A slot is considered attested if either one of the following conditions is met:

The is_proto_attested field in the content of the cell is true, and attestation_threshold_percent is None (default DAL behavior).

The attestation_threshold_percent is set to a specific threshold, the slot meets or exceeds this threshold, and the publisher is authorized, provided a whitelist is specified in restricted_commitments_publishers.

If the slot of the page is attested, the data is retrieved from the DAL node and returned. Otherwise, the function returns None. It also returns None if the page_id or slot_id is invalid (or no longer valid). This includes cases where:

Indices are out of range.

The rollup was originated after the slot's publication.

The slot is too old (i.e., its level is beyond dal_attested_slots_validity_lag).