Module Wasm_2_0_0_pvm.Make_durable_state

Parameters

module T : Tezos_tree_encoding.TREE with type tree = Tezos_layer2_irmin_context.Irmin_context.tree

Signature

type state = T.tree

val value_length : state -> string -> int64 option Lwt.t

value_length state key returns the length of data stored for the key in the durable storage of the PVM state state, if any.

val lookup : state -> string -> bytes option Lwt.t

lookup state key returns the data stored for the key in the durable storage of the PVM state state, if any.

val list : state -> string -> string list Lwt.t

subtrees state key returns subtrees for the key in the durable storage of the PVM state state. Empty list in case if path doesn't exist.

module Tree_encoding_runner : Tezos_tree_encoding.Runner.S with type tree = state